% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_functions.R
\name{minter_colorplot}
\alias{minter_colorplot}
\title{Show Minter Theme Colors}
\usage{
minter_colorplot()
}
\value{
A plot of SI colors
}
\description{
Shows the Minter theme colors for reference.
}
\examples{
minter_colorplot()
}
