% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmarkdown_functions.R
\name{minter_draft_full_report}
\alias{minter_draft_full_report}
\title{Draft a minterbrand Report}
\usage{
minter_draft_full_report(file)
}
\arguments{
\item{file}{File name for the draft}
}
\description{
Drafts a new minterbrand full rmarkdown report
}
\examples{
minter_draft_full_report("My presentation.Rmd")
}
