% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmarkdown_functions.R
\name{minter_draft_html_report}
\alias{minter_draft_html_report}
\title{Draft a minterbrand HTML Report}
\usage{
minter_draft_html_report(file)
}
\arguments{
\item{file}{File name for the draft}
}
\description{
Drafts a new minterbrand HTMLrmarkdown report
}
\examples{
minter_draft_html_report("myReport.Rmd")
}
