% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmarkdown_functions.R
\name{minter_draft_pdf_report}
\alias{minter_draft_pdf_report}
\title{Draft a minterbrand PDF Report}
\usage{
minter_draft_pdf_report(file)
}
\arguments{
\item{file}{File name for the draft}
}
\description{
Drafts a new minterbrand PDF RMarkdown report
}
\examples{
minter_draft_pdf_report("myReport.Rmd")
}
