% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmarkdown_functions.R
\name{minter_draft_presentation}
\alias{minter_draft_presentation}
\title{Draft a minterbrand Presentation}
\usage{
minter_draft_presentation(file)
}
\arguments{
\item{file}{File name for the draft}
}
\description{
Drafts a new minterbrand ioslides presentation.
}
\examples{
minter_draft_presentation("My presentation.Rmd")
}
