% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_functions.R
\name{minter_ggplot_theme_update}
\alias{minter_ggplot_theme_update}
\title{Apply all Sorenson Impact ggplot themes}
\usage{
minter_ggplot_theme_update()
}
\value{
Invisibly sets SI ggplot theme values.
}
\description{
Applies all Sorenson Impact custom colors and settings for ggplot
}
\examples{
minter_ggplot_update()
}
