% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmarkdown_functions.R
\name{minter_knitr_settings}
\alias{minter_knitr_settings}
\title{Apply all SI knitr options}
\usage{
minter_knitr_settings(verbose = F)
}
\value{
Invisibly sets knitr options
}
\description{
Applies all minterbrand custom knitr options.
}
\examples{
minter_knitr_settings()
}
